backend/requirements.txt
graypy==2.1.0
# PyGPUtil==0.5.0
gputil==1.4.0
fastapi==0.110.0
uvicorn==0.30.0
pydantic==2.5.2
pydantic-settings==2.1.0
psutil==5.9.5
pynput==1.7.6
pyautogui==0.9.54
python-dotenv==1.0.0
colorama==0.4.6
pywin32==306 ; platform_system == "Windows"
pyobjc-framework-Quartz==9.2 ; platform_system == "Darwin"
pytweening==1.0.7
PyYAML==6.0.1
SQLAlchemy==2.0.25
aiofiles==23.2.1
websockets==12.0
sentry-sdk[fastapi]==2.1.1